requirements-dev.txt
# Development dependencies
-r requirements.txt

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-flask>=1.2.0

# Code quality
black>=23.0.0
flake8>=6.0.0
pylint>=2.17.0
isort>=5.12.0

# Documentation
sphinx>=6.0.0
sphinx-rtd-theme>=1.2.0

# Debugging
ipdb>=0.13.0
